+++ src/main/java/com/dev/product/Coffee/implement/ContactServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -24,26 +25,35 @@ public class ContactServiceImpl implements ContactService {
 
     @Override
     public ContactEntity createContact(ContactEntity contact) {
-        contact.setCreated_date(new Date());
+        contact.setCreatedDate(new Date());
         contactRepository.save(contact);
         return contact;
     }
+
     @Override
     public List<ContactEntity> getAllContact() {
         return contactRepository.findAll();
     }
+
     @Override
     public boolean deleteContact(Long id) {
-        ContactEntity contactEntity = contactRepository.findById(id).get();
-        contactRepository.delete(contactEntity);
-        return true;
+        Optional<ContactEntity> contactEntityOptional = contactRepository.findById(id);
+        if (contactEntityOptional.isPresent()) {
+            ContactEntity contactEntity = contactEntityOptional.get();
+            contactRepository.delete(contactEntity);
+            return true;
+        }
+        return false;
     }
 
     @Override
     public ContactEntity getContactById(Long id) {
-        ContactEntity contactEntity = contactRepository.findById(id).get();
+        Optional<ContactEntity> contactEntityOptional = contactRepository.findById(id);
+        ContactEntity contactEntity = null;
+        if (contactEntityOptional.isPresent()) {
+            contactEntity = contactEntityOptional.get();
+        }
         return contactEntity;
     }
-
 
 }
